requirements.txt
langchain-core>=0.3.0
langchain-groq>=0.1.0
langgraph>=0.2.0
python-dotenv>=1.0.0
galileo>=1.0.0
langchain-community>=0.3.0
pydantic>=2.0.0
tenacity>=8.2.0
langgraph-checkpoint>=0.1.0
ddgs>=0.1.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
requests>=2.31.0

